Fix temperature reading command
Turns out that my particular printer sends T0 as the temperature reading for the first tool. Other printers tend to use E to prefix the hotend temperature. Fixing it for now by explicitly search for the T0 reading. In the future I need to fix this to support more tools.